requirements.txt
datasets
radon
matplotlib
seaborn
pandas
numpy
scipy
transformers
torch
scikit-learn
umap-learn
pandas
tqdm
psutil
joblib